# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Ticketing API Documentation",
        "description": "Creation of an API in PHP Laravel for event ticketing",
        "contact": {
            "email": "ronsard2014@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/order-intents": {
            "get": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Get all order intents",
                "description": "Retrieve a list of order intents with pagination",
                "operationId": "getOrderIntents",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    },
                    "500": {
                        "description": "Unable to fetch events"
                    }
                }
            },
            "post": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Create a new order intent",
                "description": "Store a new order intent",
                "operationId": "StoreOrderIntents",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_intent_price",
                                    "order_intent_type",
                                    "user_email",
                                    "user_phone",
                                    "expiration_date"
                                ],
                                "properties": {
                                    "order_intent_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "order_intent_type": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "user_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "user_phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-31T23:59:59Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/order-intents/{id}": {
            "get": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Get a specific order intent",
                "description": "Retrieve a specific order intent by ID",
                "operationId": "ShowOrderIntents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order intent ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "put": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Update an existing order intent",
                "description": "Update an existing order intent by ID",
                "operationId": "UpdateOrderIntents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order intent ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_intent_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 99.99
                                    },
                                    "order_intent_type": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "user_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "user_phone": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-31T23:59:59Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated successfully"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order Intents"
                ],
                "summary": "Delete an order intent",
                "description": "Delete an existing order intent by ID",
                "operationId": "DestroyOrderIntents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order intent ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order intent deleted"
                    },
                    "404": {
                        "description": "Order intent not found"
                    }
                }
            }
        },
...
    },
...
}